pyproject.toml
[tool.poetry]
name = "f1-rag"
version = "0.1.0"
description = "RAG system dedicated to retrieving relevant information about past F1 races. Also contains a wikipedia-downloader module to retrieve all articles related to F1."
authors = ["Colin Llacer"]
readme = "README.md"
packages = [
    { include = "F1_RAG", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
wikipedia-api = "^0.7.1"
pytest = "^8.3.3"
haystack-ai = "^2.6.1"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.2.1"
flake8 = "^7.1.1"

[tool.pytest.ini_options]
pythonpath = [
    "."
]
filterwarnings = [
    "ignore::DeprecationWarning:haystack.*"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
